requirements.txt
flask==3.0.3
flask-cors==4.0.1
supabase==2.5.0
python-dotenv==1.0.1
